*** empty log message ***
git-svn-id: svn://svn.code.sf.net/p/actiongame/code/trunk/ac@1272 e016b1c5-ed68-44aa-a2f7-31cac88cffde

+++ source/src/md2.h
@@ -166,7 +166,7 @@ struct md2 : vertmodel
                 return;
             }
             int n = animfr[anim];
-            if(anim==ANIM_PAIN || anim==ANIM_DEATH || anim==ANIM_LYING_DEAD) n += varseed%3;
+            if(anim==ANIM_PAIN || anim==ANIM_DEATH || anim==ANIM_LYING_DEAD) n += uint(varseed)%3;
             as.frame = frame[n];
             as.range = range[n];
         }
